Move exception -> _exception
