+++ pyoozie/__init__.py
@@ -7,7 +7,7 @@ from pyoozie.builder import WorkflowBuilder, CoordinatorBuilder
 from pyoozie.model import parse_coordinator_id, parse_workflow_id, ArtifactType, \
     CoordinatorStatus, CoordinatorActionStatus, WorkflowStatus, WorkflowActionStatus
 from pyoozie.oozie_client import OozieClient
-from pyoozie.exceptions import OozieException
+from pyoozie._exceptions import OozieException
 
 __version__ = '0.0.0'
