+++ pyoozie/_exceptions.py
